pyproject.toml
[tool.poetry]
package-mode = false
description = ""
authors = ["Juan M. Miramont <jmiramontt@univ-lille.fr>"]
readme = "README.md"
packages = [{include = "src"}]

# Add your own dependencies here using "poetry add [dependency_name]"
[tool.poetry.dependencies]
python = ">=3.10,<3.11"
PyYAML = "^6.0"
pesq = "^0.0.4"
pystoi = "^0.4.1"
oct2py = "^5.6.0"
rpy2 = "3.5.2"
spatstat-interface = "^1.0.1"
mcsm-benchs = "^0.1.1"
librosa = "^0.10.2.post1"
gudhi = "^3.10.1"
# matlabengine = "9.13.6" # Check this is correct for your Matlab installation.
numba = "^0.61.0"

# nbformat = "4.2.0"
nbformat = "^5.9.2"
# ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
